feat: IngredientOrderProduct Model Created
I've created a model called IngredientOrderProduct. This model holds the
relationship between a product that's inside an order and an ingredient
that you want to add extra to that product

Let's say you order a peperoni pizza, but you want this pizza with extra
peperoni. This is relationship you would add a product pizza to an
order, and this model would add a relationship between that pizza that's
inside an order and the peperoni ingrient, which is the engridient that
you want some extra in your pizza